% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{cr_source.default}
\alias{cr_source.default}
\title{Visualized for CR Sources}
\usage{
\method{cr_source}{default}(clone1, clone2, ...)
}
\arguments{
\item{clone1}{First vector of sequences, each is a same length string of nucleotide.}

\item{clone2}{Second vector of sequences, same string-length as for the first vector.}

\item{...}{Any other arguments.}
}
\description{
Visualization of Two clones for their convergent recombination (CR) sources. Each sequence (nucleotides) is represented as a colored bar (red for A, yellow for G, blue for T and green for C) linked to its translated amino acid sequence by a colored line, red for the first clone and blue for the second.
}
