% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{network}
\alias{network}
\title{Sequences distance network}
\usage{
network(dataset, by, nrow, method, ...)
}
\arguments{
\item{dataset}{A matrix or a data frame includes row names which are used as the compared sequences. Data set's numeric values determine node-size.}

\item{by}{Index of column to set its values as node-size. first column is default (1).}

\item{nrow}{Number of nodes to display. Default is 1000 nodes.}

\item{method}{stringdist method to perform for distance dissimilarity calculation ("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex"). Default is "lv" (Levenshtein distance).}

\item{...}{Any additional arguments needed by the specialized methods.}
}
\description{
Computes pairwise string distances among repertoire's sequences and visualize similar pairs as connected nodes, each sized by its frequency.
}
\examples{

data <- matrix(rexp(1/2, n=4000), ncol = 4)
aa <- c('G', 'A', 'V', 'L', 'I', 'P', 'F', 'Y', 'W', 'S',
        'T', 'N', 'Q', 'C', 'M', 'D', 'E', 'H', 'K', 'R')

cons <- sample(aa, 10)

aa_list <- c()
while(length(aa_list)<nrow(data)) {
  aaseq <- cons
  index <- sample(length(aaseq), 3)
  aaseq[index] <- sample(aa, 3, replace = TRUE)
  aaseq <- paste0(aaseq, collapse = "")
  aa_list <- if (!(aaseq \%in\% aa_list)) append(aa_list, aaseq)
}

rownames(data) <- aa_list
network(data, by = 3, nrow = 500)


}
